move endpoint, use decline instead of reject.
