+++ frontend/src/components/PresentationRequestList.vue
@@ -18,6 +18,7 @@
     :sort-by="['createdAt']"
     :sort-desc="[false]"
     single-select
+    @click:row="openPresentation"
   >
     <template v-slot:[`item.createdAt`]="{ item }">
       {{ item.createdAt | formatDateLong }}
